# Backend/MyBackend/Models/PatchDiscordBody.cs
namespace MyBackend.Models
{
    public class PatchDiscordBody
    {
        public string Discord {  get; set; } = string.Empty;
        public string UserName { get; set; } = string.Empty;
    }
}