# Backend/MyBackend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MyBackend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/users/logInUser": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/registerUser": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ParametersRegisterResponse"
            }
          }
        }
      }
    },
    "/api/users/getUserGmPicture": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "gmUserName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/relogInUser": {
      "get": {
        "tags": [
          "Users"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "userName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          }
        }
      }
    },
    "/api/users/userName": {
      "patch": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PatchUserNameBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/email": {
      "patch": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PatchEmailBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/profilePicture": {
      "patch": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PatchProfilePicBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/discord": {
      "patch": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PatchDiscordBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/user": {
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ParametersRegisterResponse": {
      "type": "object",
      "properties": {
        "emailIsUnique": {
          "type": "boolean"
        },
        "emailIsCorrect": {
          "type": "boolean"
        },
        "passwordIsCorrect": {
          "type": "boolean"
        },
        "userNameIsCorrect": {
          "type": "boolean"
        }
      }
    },
...
    "PatchDiscordBody": {
      "type": "object",
      "properties": {
        "discord": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
    "PatchEmailBody": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
    "PatchProfilePicBody": {
      "type": "object",
      "properties": {
        "profilePic": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        }
      }
    },
    "PatchUserNameBody": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
    "UserModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "discord": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        }
      }
    }
  }
}